refactor(apex-lsp-compliant-services): extract worker logging helper - W-24145209
Move emitWorkerLog into src/worker/workerLog.ts. A leaf that maps a string
level onto the shared logger; imported by the shared module and, in later
slices, by the resolution substrate and rename core.

W-24145209
